lilypond.make: Call `rsync-test.py' with `--gub-target-dir' option.
Note that this commit needs lilypond issue #5456 so that relative paths are
supported in the `output-distance.py' script.  For testing older versions of
lilypond, you have to undo this commit.

Barring errors in this series of patches, gub can now do a complete lilypond
build (including regression tests and documention) that is no longer bound
to a certain user, top-level directory, platform, or architecture.