Enable compiling librestrict on macOS.
librestrict (source in the top-level directory of GUB) uses C aliases.
Darwin (the underlying operating system that macOS is built on) does
not support aliases, thus librestrict can not currently be compiled on
macOS. This commit defines a 'safe_alias' macro that smooths over this
missing feature by providing the current behavion on non-darwin systems
and providing a workaround on darwin systems. The macro is defined in
the 'alias.h' file inside the librestrict directory.