Avoid changing the timestamp of config.make
This generalizes the effect of an existing hack.  Before, the scripts
xetex-with-options and xelatex-with-options were touched to avoid
having them regenerated (i.e. overwritten) as a result of the
modification of config.make.  Now, *no* target having config.make as a
prerequisite will consider it modified.

The issue motivating this change is that modification to config.make
triggers removal of fonts in out/..., and for some reason they are not
recreated before they are needed in the gub environment.  Attempts to
reproduce this issue by touching the involved files in a regular
lilypond build environment have been unsuccessful: there, the fonts do
exist in out/... after `make test` is run.